% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_curvePlot.R
\name{ffaPlot}
\alias{ffaPlot}
\title{Plot Flood Frequency Analysis Curve}
\usage{
ffaPlot(plotPositions, LP3fit, outFile, siteNum = "", siteName = "", PILF = 0)
}
\arguments{
\item{plotPositions}{dataframe containing observed peak values in a column named "peak_va",corresponding plotting positions in a column named "plotPosition",
and an optional column named "historic" containing a logical TRUE for peaks which are historic}

\item{LP3fit}{dataframe containing LP3 curve fit for a site with exceedance probabilities in a column named "EXC_Prob", and corresponding peak estimates in a column named "Estimate",
lower confidence limits in a column named "Conf_Low", and upper confidence limits in a column named "Conf_Up"}

\item{outFile}{file path to output graph}

\item{siteNum}{character - site number to print on plot}

\item{siteName}{character - site name to print on plot}

\item{PILF}{numeric - PILF threshold for analysis}
}
\description{
Function to plot results of peak-flow frequency analysis
}
