% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectGit.R
\name{new.function}
\alias{new.function}
\title{Title - print squares of numbers in sequence}
\usage{
new.function(a)
}
\arguments{
\item{a}{Any number}
}
\value{
Squares of numbers 1 through a in sequence
}
\description{
Title - print squares of numbers in sequence
}
